Remove `shift`/`reset`.
The problem with native delimited continuations is that they give
every block an implicit sequential ordering, which harms parallelism.

If you receive a block from somewhere, you can't just split it in half
and start evaluating the parts, because the leftmost half is able to
view and suppress the computation of the rightmost half with `shift`.

I still think delimited continuations are important, but this sort of
continuation passing style has to be explicit and opt-in. When a block
is in CPS it will be obvious because of the way it's "folded" in on
itself, while other blocks will be able to enjoy ambient parallelism.